requirements.txt
# Core Django & Web Framework
Django==5.2.6
djangorestframework==3.15.2
gunicorn==23.0.0
whitenoise
asgiref==3.9.1

# Django Extensions
django-cors-headers==4.6.0
django-filter==24.3
django-health-check==3.20.0
django-imagekit==5.0.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-webpack-loader==3.0.0
django-widget-tweaks==1.5.0
django-appconf==1.1.0

# Database
psycopg2-binary==2.9.10
pgvector==0.2.1

# Celery (Background Tasks)
celery==5.4.0
redis==5.1.1
amqp==5.3.1
billiard==4.2.2
kombu==5.5.4
vine==5.1.0

# AI & ML - Core
torch==2.8.0
transformers==4.56.2
sentence-transformers==3.0.1
pillow==11.0.0
numpy==2.3.2
safetensors==0.6.2
huggingface-hub==0.35.0
tokenizers==0.22.1

# AI & ML - Dependencies
scikit-learn==1.7.2
scipy==1.16.2
nltk==3.9.1
joblib==1.5.2
regex==2025.9.18

# NVIDIA GPU Support (for future GPU processing)
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
triton==3.4.0

# Video Processing
moviepy==1.0.3
imageio==2.37.0
imageio-ffmpeg==0.6.0
proglog==0.1.12

# Image Processing
pilkit==3.0
networkx==3.5

# Utilities
python-decouple==3.8
python-dotenv==1.1.1
python-magic==0.4.27
requests==2.32.5
beautifulsoup4==4.13.5
certifi==2025.8.3
python-dateutil==2.9.0.post0
tzdata==2025.2
urllib3==2.5.0

# Additional Dependencies
annotated-types==0.7.0
anyio==4.11.0
charset-normalizer==3.4.3
cffi==1.2.0
cryptography==46.0.1
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
psutil==7.1.0
pycparser==2.23
Pygments==2.19.2
PyYAML==6.0.2
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typing_extensions==4.15.0
wcwidth==0.2.14
setuptools==80.9.0
Markdown==3.7
